requirements.txt
fastapi==0.85.0
pymongo==4.2.0
tenacity==8.1.0
uvicorn==0.18.3
python-dotenv==0.21.0
python-jose==3.3.0
passlib==1.7.4

